requirements.dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.dev.txt requirements.dev.in
#
astroid==1.6.0            # via pylint
click==6.7                # via pip-tools
first==2.0.1              # via pip-tools
isort==4.2.15             # via pylint
lazy-object-proxy==1.3.1  # via astroid
mccabe==0.6.1             # via pylint
pip-tools==1.11.0
pylint==1.8.1
pyyaml==3.12
six==1.11.0               # via astroid, pip-tools, pylint
termcolor==1.1.0
wrapt==1.10.11            # via astroid
